Сделал экрна с платежами

+++ app/src/main/java/com/martynov/testaeon/dto/Payments.kt
@@ -4,5 +4,5 @@ data class Payments(
         val desc: String,
         val amount: Double,
         val currency: String,
-        val created: Int
+        val created: Long
 )
